move common LR functions to portal view, fix if checks
