+++ portal/eproms/templates/eproms/resources.html
